+++ pages/car-models/index.vue
@@ -4,6 +4,23 @@
       <h1 class="title">
         Modèles de véhicules
       </h1>
+      <button
+        class="button is-rounded"
+        type="button"
+        @click="toggleCsvModal(true)"
+      >
+        <fa-icon
+          :icon="['fas', 'file-export']"
+          class="has-text-info"
+        />
+        Exporter CSV
+      </button>
+      <csv-modal
+        :csv-status="displayModal"
+        :pagination="pagination"
+        :api-call="$api.carModels.getCarModels"
+        @toggleModal="toggleCsvModal"
+      />
       <div class="options">
         <nuxt-link
           :to="{name: 'car-models-new'}"
@@ -54,6 +71,7 @@
 
 <script>
 import ecList from '~/components/crud/list.vue';
+import csvModal from '~/components/csv-modal';
 
 const columns = { id: 'ID', label: 'Label' };
 
@@ -65,6 +83,7 @@ export default {
   watchQuery: ['offset', 'limit'],
   components: {
     ecList,
+    csvModal,
   },
   async asyncData({ $api, query }) {
     const offset = parseInt(query.offset, 10) || 0;
@@ -73,6 +92,11 @@ export default {
     return {
       carModels: data,
       pagination,
+    };
+  },
+  data() {
+    return {
+      displayModal: false,
     };
   },
   computed: {
@@ -86,6 +110,9 @@ export default {
       const updatedList = await getCarModels.call(this.$api, offset, limit);
       this.carModels = updatedList.data;
       this.pagination = updatedList.pagination;
+    },
+    toggleCsvModal(force) {
+      this.displayModal = force || !this.displayModal;
     },
   },
 };
