+++ pages/users/index.vue
@@ -4,6 +4,23 @@
       <h1 class="title">
         Utilisateurs
       </h1>
+      <button
+        class="button is-rounded"
+        type="button"
+        @click="toggleCsvModal(true)"
+      >
+        <fa-icon
+          :icon="['fas', 'file-export']"
+          class="has-text-info"
+        />
+        Exporter CSV
+      </button>
+      <csv-modal
+        :csvStatus="displayModal"
+        @toggleModal="toggleCsvModal"
+        :pagination="pagination"
+        :apiCall="$api.users.getUsers"
+      />
       <div class="options">
         <nuxt-link
           v-if="$auth.isRegulator()"
@@ -54,20 +71,27 @@
 </template>
 
 <script>
-import ecList from '~/components/crud/list.vue';
+import ecList from '~/components/crud/list';
+import csvModal from '~/components/csv-modal';
 
 export default {
   watchQuery: ['offset', 'limit'],
   components: {
     ecList,
+    csvModal,
   },
   async asyncData({ $api, query }) {
     const offset = parseInt(query.offset, 10) || 0;
     const limit = parseInt(query.limit, 10) || 30;
-    const { data, pagination } = await $api.users.getUsers('id,email', offset, limit);
+    const { data, pagination } = await $api.users.getUsers('id,email', { offset, limit });
     return {
       users: data,
       pagination,
+    };
+  },
+  data() {
+    return {
+      displayModal: false,
     };
   },
   methods: {
@@ -75,9 +99,12 @@ export default {
       await this.$api.users.deleteUser(id);
       const offset = parseInt(this.$route.query.offset, 10) || 0;
       const limit = parseInt(this.$route.query.limit, 10) || 30;
-      const updatedList = await this.$api.users.getUsers('id,email', offset, limit);
+      const updatedList = await this.$api.users.getUsers('id,email', { offset, limit });
       this.users = updatedList.data;
       this.pagination = updatedList.pagination;
+    },
+    toggleCsvModal(force) {
+      this.displayModal = force || !this.displayModal;
     },
   },
 };
